% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasaAsteroid.R
\docType{class}
\name{nasaAsteroid-class}
\alias{nasaAsteroid-class}
\alias{nasaAsteroid}
\title{Nasa Asteroid Data Analysis}
\description{
Data analysis on nasa asteroid data source NASA Open Api
}
\section{Fields}{

\describe{
\item{\code{formula}}{Formula}
}}

\section{Methods}{

\describe{
\item{\code{fittedValues()}}{Fitted Values calculation}

\item{\code{getAsteroidsAsDataFrame()}}{Return asteroids as data frame}

\item{\code{hazardousAsteroids()}}{Calculates how many asteroids are there and how many of them are hazardous}

\item{\code{meanSummary()}}{Calculates mean for some columns}

\item{\code{medianSummary()}}{Calculates median for some columns}

\item{\code{regressionCoefficients()}}{Regression Coefficient calculation}

\item{\code{residuals()}}{Residuals calculation}
}}

