% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasaAsteroidPackage.R
\docType{package}
\name{nasaAsteroidPackage}
\alias{nasaAsteroidPackage}
\alias{nasaAsteroidPackage-package}
\title{A package for analyzing Asteroids data from NASA Open API}
\description{
This package calculates some values for asteroids
}
\author{
Fahad Hameed and Ugurcan Lacin
}
