% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasaAsteroid.R
\docType{class}
\name{nasaAsteroid-class}
\alias{nasaAsteroid-class}
\alias{nasaAsteroid}
\title{Nasa Asteroid Data Analysis}
\description{
You can have data analysis on Asteroids data from NASA Open API
}
\section{Fields}{

\describe{
\item{\code{formula}}{Formula}
}}


